Move api to middleware package root
