+++ api/app/controllers/tags.py
@@ -2,9 +2,10 @@
 # -*- coding: utf-8 -*-
 
 from flask import request
-from flask_restplus import Resource
+from flask_restplus import Resource, fields
 from sqlalchemy import func
 
+from middleware import api
 from middleware.cache import cache_filler, is_cache_filler
 from middleware.json_response import json_response
 from models import Tag, Thesis
@@ -15,8 +16,9 @@ from services.logger import logger
 class TagsView(Resource):
     decorators = [cache_filler(), cache.cached()]
 
+    @api.doc(params={"filename": "If set, return a downloadable file with this name"})
     def get(self, filename=None):
-        """List of all known tags."""
+        """List all tags."""
 
         if not is_cache_filler():
             logger.info("Cache miss for {}".format(request.path))
@@ -45,9 +47,19 @@ class TagsView(Resource):
         return json_response(rv, filename=filename)
 
 
+auth = api.model('auth', {
+    'admin_key': fields.String(required=True, description="Admin auth key")
+})
+
+
+delete_request = api.parser()
+delete_request.add_argument("admin_key", type=str, location="body")
+
+
 class TagView(Resource):
     decorators = [cache_filler(), cache.cached()]
 
+    @api.doc(params={"slug": "Slug for the requested tag"})
     def get(self, slug: str):
         """Return data for all theses in a tag."""
         if not is_cache_filler():
@@ -59,7 +71,7 @@ class TagView(Resource):
             return json_response({"error": "Tag not found"}, status=404)
 
         rv = {
-            "data": tag.to_dict(include_related_tags='full'),
+            "data": tag.to_dict(include_related_tags="full"),
             "theses": [thesis.to_dict() for thesis in tag.theses],
             "elections": {
                 thesis.election_id: thesis.election.to_dict() for thesis in tag.theses
@@ -68,6 +80,7 @@ class TagView(Resource):
 
         return json_response(rv)
 
+    @api.expect(auth, validate=True)
     def delete(self):
         admin_key = request.get_json().get("admin_key", "")
         if admin_key != app.config.get("ADMIN_KEY"):
@@ -84,7 +97,7 @@ class TagView(Resource):
         db.session.commit()
 
         rv = {
-            "data": tag.to_dict(include_related_tags='full'),
+            "data": tag.to_dict(include_related_tags="full"),
             "theses": [thesis.to_dict() for thesis in tag.theses],
             "elections": {
                 thesis.election_id: thesis.election.to_dict() for thesis in tag.theses
